rename files and schema class
